token: implement EAR token generation
This commit allows the AS to issue EAR tokens with the
help of the rust-ear crate.

EAR tokens require particular claims. This creates a binding
between the AS policy and the EAR token.
Specifically, the policy engine must return an EAR appraisal.
The policy engine is still generic. Multiple policy engines
could be implemented as long as they create an appraisal.

Token generation is no longer generic.
Since the policy engine, will always return an appraisal,
we must generate an EAR token.
This commit removes the simple token issuer
and replaces the TokenProvider trait with a struct.

The KBS will still be able to validate many different tokens,
but this commit changes the AS to only issue EAR tokens.

There are a few other changes, including that the policy engine
no longer takes multiple policies. For now, we only evaluate
the first policy in the policy list, but future commits will
change this convention so that we only ever think about one
policy for the attestation service (until we introduce support
for validating multiple devices at once).

This commit also removes the flattening of the tcb claims.
With the EAR tokens, we store the TEE pubkey the tcb claims.
If these claims are flattened, we will need to do some extra
work to deserialize the key.

The TCB claims are currently flattened so that we can use the
key names as the input to the RVPS. This commit breaks this
functionality, but a future commit will change the way
the RVPS works to accomodate. There isn't a direct pairing
between claim names and reference values, so there is no
reason to keep flattening all the claims, especially
because the flattening code has some corner cases
that it does not support.

This commit also adds the init_data_claims and
runtime_data_claims to the tcb claims as long as
the corresponding claims about the hashes are already there.
This will allow the init_data to travel with the token,
which will be convenient except if the init_data is too big.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>