docker-compose: add attestation token keypair
A keypair is required to sign and validate the attestation token.
In the past this was optional, but now it is not.

Update the docker-compose manifest and configs to pass in this
new keypair and update the docs to tell people how to generate it.

This does complicate the user experience, but things are not secure
without it. That said, we may be able to implement this automatically
in a future PR.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>