protocols: change policy_ids to policy_id
When generating EAR tokens, it seems best to only use one policy
at a time (per-submod).

In the commit that introduces EAR token generation in the AS,
we simply ignore all policies in the policy_ids list except
the first one.

Here, we change the interface so that only one policy can be
provided in an attestation request

The KBS always sets one policy ("default"), anyway.
In the future, we should figure out how to set this policy id
more dynamically.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>