policy: decouple policy claims from policy engine
If we want to support multiple types of tokens,
we'll need to decouple the claims that are evaluated in the policy
from the poliicy engine.

This allows the token broker to specify a set of rules that
the policy engine will evaluate. Then, the token broker
will get the unprocessed output from the policy engine
and do what it needs to.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>